# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Akamai CCM API
  version: 1.0.0
  description: >
    Cloud Compliance Manager API for managing Linode accounts, resources,
    compliance rules, reports, and users.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://<your-host>
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          type: string
          enum: [admin, power_user, auditor]
        can_view_costs:
          type: boolean
        can_view_compliance:
          type: boolean
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
    Account:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        api_token:
          type: string
        webhook_api_key:
          type: string
          nullable: true
        last_sync_at:
          type: string
          format: date-time
          nullable: true
        last_evaluated_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        detail:
          type: string
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register first admin user
      description: Only succeeds when no users exist in the system.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                full_name:
                  type: string
                  default: ''
      responses:
        '200':
          description: Registered and authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Registration is closed (users already exist)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      responses:
        '200':
          description: Current authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ─── Accounts ─────────────────────────────────────────────────────────────
  /api/accounts:
    get:
      tags: [Accounts]
      summary: List accounts
      responses:
        '200':
          description: List of accessible accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
    post:
      tags: [Accounts]
      summary: Create account
      description: Requires `power_user` or `admin` role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, api_token]
              properties:
                name:
                  type: string
                api_token:
                  type: string
                webhook_api_key:
                  type: string
      responses:
        '200':
          description: Created account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/accounts/{account_id}:
    parameters:
      - name: account_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Accounts]
      summary: Get account details
      responses:
        '200':
          description: Account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [Accounts]
      summary: Update account
      description: Requires `power_user` or `admin` role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                api_token:
                  type: string
                webhook_api_key:
                  type: string
      responses:
        '200':
          description: Updated account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
    delete:
      tags: [Accounts]
      summary: Delete account
      description: Requires `admin` role.
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean

  # ─── Resources ────────────────────────────────────────────────────────────
...